refactor: BonusNumber 객체 생성
세부사항
- LottoNumber를 상속받는 BonusNumber 객체 생성
- 당첨 번호에 보너스 볼이 포함되어 있는지 검증을 BonusNumber에서 처리
