+++ src/main/java/controller/LottoController.java
@@ -1,5 +1,6 @@
 package controller;
 
+import model.BonusNumber;
 import model.Lotto;
 import model.LottoGenerator;
 import model.LottoNumber;
@@ -25,7 +26,7 @@ public class LottoController {
         OutputView.showLotto(transToLottosDto(lottos), lottos.getBuyLottoCount());
 
         final Lotto winningLotto = Lotto.fromStringsInput(InputView.inputWinningLotto());
-        final LottoNumber bonusNumber = LottoNumber.forBonusNumber(InputView.inputBonusNumber(), transToLottoDto(winningLotto));
+        final LottoNumber bonusNumber = BonusNumber.of(InputView.inputBonusNumber(), transToLottoDto(winningLotto));
 
         final LottoResult result = lottos.getResult(winningLotto, bonusNumber);
         OutputView.showResult(result.getResult(), result.getRateOfReturn(lottoPurchaseMoney.getValue()));
