+++ src/main/java/model/LottoNumber.java
@@ -1,6 +1,5 @@
 package model;
 
-import java.util.List;
 import java.util.Objects;
 import java.util.regex.Pattern;
 
@@ -22,28 +21,15 @@ public class LottoNumber {
         return new LottoNumber(Integer.parseInt(input));
     }
 
-    public static LottoNumber forBonusNumber(final String input, final List<Integer> winningLotto) {
-        validateStringInput(input);
-        int number = Integer.parseInt(input);
-        validateContainInput(number, winningLotto);
-        return new LottoNumber(number);
-    }
-
     private void validateNumber1to45(final int number) {
         if (number < LOTTO_MIN_NUMBER || number > LOTTO_MAX_NUMBER) {
             throw new IllegalArgumentException("로또 번호 및 보너스 볼은 1과 45사이의 숫자이어야 합니다.");
         }
     }
 
-    private static void validateStringInput(final String input) {
+    protected static void validateStringInput(final String input) {
         if (!REGEX.matcher(input).matches()) {
             throw new IllegalArgumentException("숫자입력만 허용합니다.");
-        }
-    }
-
-    private static void validateContainInput(int input, List<Integer> lotto) {
-        if (lotto.contains(input)) {
-            throw new IllegalArgumentException("보너스 볼은 당첨 번호에 포함되어 있으면 안됩니다.");
         }
     }
