+++ src/test/java/model/BonusNumberTest.java
@@ -1,0 +1,26 @@
+package model;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class BonusNumberTest {
+
+    @DisplayName("보너스 볼이 당첨 번호에 포함되어 있으면 예외를 발생한다.")
+    @ValueSource(strings = {"1", "2", "3", "4", "5", "6"})
+    @ParameterizedTest
+    void bonus_number_in_winning_lotto(String input) {
+        // given
+        List<Integer> winningLotto = List.of(1, 2, 3, 4, 5, 6);
+
+        // when
+        // then
+        assertThatThrownBy(() -> BonusNumber.of(input, winningLotto))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("보너스 볼은 당첨 번호에 포함되어 있으면 안됩니다.");
+    }
+}
